third_party/WebKit/LayoutTests/media/track/track-word-breaking.html

<!DOCTYPE html>
<title>Test that line breaks are forced in captions.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<script>
// This should not be necessary, but due to crbug.com/372245 style
// gets dirtied during layout in LayoutTextTrackContainer which
// causes assertions to fail when Document::scrollingElement is
// queried during compositing update (a valid request).
window.internals.settings.setCompositorWorkerEnabled(false);
</script>
<video>
    <track src="captions-webvtt/long-word.vtt" kind="captions" default>
</video>
<script>
async_test(function(t) {
    var singleLineCaptionHeight;
    var video = document.querySelector("video");
    video.oncanplaythrough = t.step_func(function() {
        // Record Single line caption height.
        singleLineCaptionHeight = getCaptionLineHeight();
        video.currentTime = 3;
    });

    video.onseeked = t.step_func_done(function() {
        var multiLineCaptionHeight = getCaptionLineHeight();
        assert_greater_than(multiLineCaptionHeight / singleLineCaptionHeight, 2);
    });

    video.src = findMediaFile("video", "../content/test");

    function getCaptionLineHeight() {
        var captionElement = textTrackContainerElement(video).firstChild.firstChild;
        return captionElement.getBoundingClientRect().height;
    }
});
</script>
